pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "cfsonarrmatcher"
version = "0.2.7"
dependencies = [ "rapidfuzz", "python-dateutil", "fauxlogger @ git+https://github.com/codefaux/python-fauxlogger@main" ]